adding backend basic auth enforcer implementation
